requirements.txt
# Core ML Libraries
torch>=2.1.0
transformers>=4.36.0
datasets>=2.16.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Training & Optimization
accelerate>=0.25.0
trl>=0.7.0
sentencepiece>=0.1.99

# Evaluation
sqlparse>=0.4.4
func-timeout>=4.3.5

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0

# Logging & Monitoring
wandb>=0.16.0
tensorboard>=2.15.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
requests>=2.31.0
